requirements.txt
# Core ML
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# NLP
transformers>=4.30.0
nltk>=3.8.0
spacy>=3.6.0

# MLOps
mlflow>=2.5.0
optuna>=3.3.0
optuna-dashboard>=0.13.0

# API
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Utils
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.65.0

# Development
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.4.0
isort>=5.12.0
pre-commit>=3.3.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.17.1,<7.0.0
ipywidgets>=8.1.0

